Refactor public API to work with keyset
