+++ lib/graphql/connections/stable.rb
@@ -11,66 +11,17 @@ module GraphQL
     #
     # For more information see GraphQL Cursor Connections Specification
     # https://relay.dev/graphql/connections.htm
-    class Stable < ::GraphQL::Connections::Base
-      def initialize(*args, primary_key: nil, **kwargs)
-        @primary_key = primary_key
-
-        super(*args, **kwargs)
-      end
-
-      def has_previous_page # rubocop:disable Naming/PredicateName
-        if last
-          nodes.any? && items.where(arel_table[primary_key].lt(nodes.first[primary_key])).exists?
-        elsif after
-          items.where(arel_table[primary_key].lteq(after_cursor)).exists?
-        else
-          false
-        end
-      end
-
-      def has_next_page # rubocop:disable Naming/PredicateName
-        if first
-          nodes.any? && items.where(arel_table[primary_key].gt(nodes.last[primary_key])).exists?
-        elsif before
-          items.where(arel_table[primary_key].gteq(before_cursor)).exists?
-        else
-          false
-        end
-      end
-
-      def cursor_for(item)
-        cursor = serialize(item[primary_key])
-        cursor = encode(cursor) if opaque_cursor
-        cursor
-      end
-
-      private
-
-      def limited_relation
-        scope = sliced_relation
-        nodes = []
-
-        if first
-          nodes |= scope.
-                   reorder(arel_table[primary_key].asc).
-                   limit(first).
-                   to_a
+    class Stable
+      def self.new(*args, desc: false, keys: nil, **kwargs)
+        if kwargs[:primary_key] || keys.nil?
+          raise NotImplementedError, 'desc connection is not implemented yet' if desc
+          return GraphQL::Connections::Key.new(*args, **kwargs)
         end
 
-        if last
-          nodes |= scope.
-                   reorder(arel_table[primary_key].desc).
-                   limit(last).
-                   to_a.reverse!
-        end
+        raise ArgumentError, 'keyset for more that 2 keys is not implemented yet' if keys.length > 2
 
-        nodes
-      end
-
-      def sliced_relation
-        items.
-          yield_self { |s| after ? s.where(arel_table[primary_key].gt(after_cursor)) : s }.
-          yield_self { |s| before ? s.where(arel_table[primary_key].lt(before_cursor)) : s }
+        klass = desc ? GraphQL::Connections::Keyset::Desc : GraphQL::Connections::Keyset::Asc
+        klass.new(*args, **kwargs)
       end
     end
   end
